move fake data source and fix test cases
